examples/chatbot_app/requirements.txt
streamlit>=1.32.0
snowflake-snowpark-python>=1.11.0
